refactor(data): rename src/plugins/async → src/plugins/data to match public export
