+++ src/index.ts
@@ -29,8 +29,8 @@ export { autosize } from "./plugins/autosize";
 export type { AutosizePluginConfig } from "./plugins/autosize";
 export { masonry } from "./plugins/masonry";
 export type { MasonryPluginConfig } from "./plugins/masonry";
-export { data } from "./plugins/async";
-export type { DataPluginConfig } from "./plugins/async";
+export { data } from "./plugins/data";
+export type { DataPluginConfig } from "./plugins/data";
 export { groups } from "./plugins/groups";
 export type { GroupsPluginConfig } from "./plugins/groups";
 export { table } from "./plugins/table";
