+++ test/plugins/data/plugin.test.ts
@@ -8,7 +8,7 @@
 
 import { describe, it, expect, beforeAll, afterAll, mock } from "bun:test";
 import { GlobalRegistrator } from "@happy-dom/global-registrator";
-import { data as dataPlugin } from "../../../src/plugins/async/plugin";
+import { data as dataPlugin } from "../../../src/plugins/data/plugin";
 import type { VListItem, VListAdapter } from "../../../src/types";
 import { createPluginMockContext } from "../../helpers/plugin-context";
 import { createEmitter } from "../../../src/events/emitter";
